# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MealType = (typeof MealType)[keyof typeof MealType];

export const MealType = {
  breakfast: "breakfast",
  lunch: "lunch",
  dinner: "dinner",
} as const;

export interface Meal {
  id: number;
  name: string;
  description?: string | null;
  scheduledDate?: string | null;
  mealType?: MealType | null;
  position: number;
  createdAt: string;
}

export interface CreateMealBody {
  name: string;
  description?: string | null;
  scheduledDate?: string | null;
  mealType?: MealType | null;
}

export interface UpdateMealBody {
  name?: string;
  description?: string | null;
  scheduledDate?: string | null;
  mealType?: MealType | null;
  position?: number;
}

export interface MoveMealBody {
  scheduledDate?: string | null;
  mealType?: MealType | null;
  position: number;
}

export interface Day {
  date: string;
  dayOfWeek: string;
  isToday: boolean;
}

export interface Ingredient {
  id: number;
  mealId: number;
  name: string;
  measurement?: string | null;
  createdAt: string;
}

export interface CreateIngredientBody {
  name: string;
  measurement?: string | null;
}

export interface MealWithIngredients {
  id: number;
  name: string;
  description?: string | null;
  scheduledDate?: string | null;
  mealType?: MealType | null;
  position: number;
  createdAt: string;
  ingredients: Ingredient[];
}

export type ListDaysParams = {
  /**
   * Start date (defaults to today)
   */
  startDate?: string;
  /**
   * Number of days to return
   */
  count?: number;
};