# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: meals
    description: Meal operations
  - name: ingredients
    description: Ingredient operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /meals:
    get:
      operationId: listMeals
      tags: [meals]
      summary: List all meals
      description: Returns all meals, both scheduled and unscheduled
      responses:
        "200":
          description: List of meals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meal"
    post:
      operationId: createMeal
      tags: [meals]
      summary: Create a new meal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMealBody"
      responses:
        "201":
          description: Created meal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
  /meals/{id}:
    put:
      operationId: updateMeal
      tags: [meals]
      summary: Update a meal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMealBody"
      responses:
        "200":
          description: Updated meal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
        "404":
          description: Meal not found
    delete:
      operationId: deleteMeal
      tags: [meals]
      summary: Delete a meal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Meal not found
  /meals/past:
    get:
      operationId: listPastMeals
      tags: [meals]
      summary: List past meals with ingredients (deduplicated by name)
      description: Returns meals scheduled before today, deduplicated by name, with their ingredients
      responses:
        "200":
          description: List of past meals with ingredients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MealWithIngredients"
  /days:
    get:
      operationId: listDays
      tags: [meals]
      summary: List days relative to a start date
      parameters:
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Start date (defaults to today)
        - name: count
          in: query
          required: false
          schema:
            type: integer
            default: 7
          description: Number of days to return
      responses:
        "200":
          description: List of day objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Day"
  /meals/{id}/ingredients:
    get:
      operationId: listIngredients
      tags: [ingredients]
      summary: List ingredients for a meal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of ingredients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ingredient"
    post:
      operationId: createIngredient
      tags: [ingredients]
      summary: Add an ingredient to a meal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIngredientBody"
      responses:
        "201":
          description: Created ingredient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
        "404":
          description: Meal not found
  /meals/{id}/ingredients/{ingredientId}:
    delete:
      operationId: deleteIngredient
      tags: [ingredients]
      summary: Delete an ingredient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: ingredientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Ingredient not found
  /meals/{id}/move:
    patch:
      operationId: moveMeal
      tags: [meals]
      summary: Move a meal to a different day/slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveMealBody"
      responses:
        "200":
          description: Moved meal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
        "404":
          description: Meal not found
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    MealType:
      type: string
      enum:
        - breakfast
        - lunch
        - dinner
    Meal:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        scheduledDate:
          type: string
          format: date
          nullable: true
        mealType:
          $ref: "#/components/schemas/MealType"
          nullable: true
        position:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - position
        - createdAt
    CreateMealBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        scheduledDate:
          type: string
          format: date
          nullable: true
        mealType:
          $ref: "#/components/schemas/MealType"
          nullable: true
      required:
        - name
    UpdateMealBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        scheduledDate:
          type: string
          format: date
          nullable: true
        mealType:
          $ref: "#/components/schemas/MealType"
          nullable: true
        position:
          type: integer
    MoveMealBody:
      type: object
      properties:
        scheduledDate:
          type: string
          format: date
          nullable: true
        mealType:
          $ref: "#/components/schemas/MealType"
          nullable: true
        position:
          type: integer
      required:
        - position
    Day:
      type: object
      properties:
        date:
          type: string
          format: date
        dayOfWeek:
          type: string
        isToday:
          type: boolean
      required:
        - date
        - dayOfWeek
        - isToday
    Ingredient:
      type: object
      properties:
        id:
          type: integer
        mealId:
          type: integer
        name:
          type: string
        measurement:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - mealId
        - name
        - createdAt
    CreateIngredientBody:
      type: object
      properties:
        name:
          type: string
        measurement:
          type: string
          nullable: true
      required:
        - name
    MealWithIngredients:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        scheduledDate:
          type: string
          format: date
          nullable: true
        mealType:
          $ref: "#/components/schemas/MealType"
          nullable: true
        position:
          type: integer
        createdAt:
          type: string
          format: date-time
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/Ingredient"
      required:
        - id
        - name
        - position
        - createdAt
        - ingredients