# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all meals, both scheduled and unscheduled
 * @summary List all meals
 */
export const ListMealsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  scheduledDate: zod.string().date().nullish(),
  mealType: zod.enum(["breakfast", "lunch", "dinner"]).nullish(),
  position: zod.number(),
  createdAt: zod.string().datetime({}),
});
export const ListMealsResponse = zod.array(ListMealsResponseItem);

/**
 * @summary Create a new meal
 */
export const CreateMealBody = zod.object({
  name: zod.string(),
  description: zod.string().nullish(),
  scheduledDate: zod.string().date().nullish(),
  mealType: zod.enum(["breakfast", "lunch", "dinner"]).nullish(),
});

/**
 * @summary Update a meal
 */
export const UpdateMealParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateMealBody = zod.object({
  name: zod.string().optional(),
  description: zod.string().nullish(),
  scheduledDate: zod.string().date().nullish(),
  mealType: zod.enum(["breakfast", "lunch", "dinner"]).nullish(),
  position: zod.number().optional(),
});

export const UpdateMealResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  scheduledDate: zod.string().date().nullish(),
  mealType: zod.enum(["breakfast", "lunch", "dinner"]).nullish(),
  position: zod.number(),
  createdAt: zod.string().datetime({}),
});

/**
 * @summary Delete a meal
 */
export const DeleteMealParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * Returns meals scheduled before today, deduplicated by name, with their ingredients
 * @summary List past meals with ingredients (deduplicated by name)
 */
export const ListPastMealsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  scheduledDate: zod.string().date().nullish(),
  mealType: zod.enum(["breakfast", "lunch", "dinner"]).nullish(),
  position: zod.number(),
  createdAt: zod.string().datetime({}),
  ingredients: zod.array(
    zod.object({
      id: zod.number(),
      mealId: zod.number(),
      name: zod.string(),
      measurement: zod.string().nullish(),
      createdAt: zod.string().datetime({}),
    }),
  ),
});
export const ListPastMealsResponse = zod.array(ListPastMealsResponseItem);

/**
 * @summary List days relative to a start date
 */
export const listDaysQueryCountDefault = 7;

export const ListDaysQueryParams = zod.object({
  startDate: zod.coerce
    .string()
    .date()
    .optional()
    .describe("Start date (defaults to today)"),
  count: zod.coerce
    .number()
    .default(listDaysQueryCountDefault)
    .describe("Number of days to return"),
});

export const ListDaysResponseItem = zod.object({
  date: zod.string().date(),
  dayOfWeek: zod.string(),
  isToday: zod.boolean(),
});
export const ListDaysResponse = zod.array(ListDaysResponseItem);

/**
 * @summary List ingredients for a meal
 */
export const ListIngredientsParams = zod.object({
  id: zod.coerce.number(),
});

export const ListIngredientsResponseItem = zod.object({
  id: zod.number(),
  mealId: zod.number(),
  name: zod.string(),
  measurement: zod.string().nullish(),
  createdAt: zod.string().datetime({}),
});
export const ListIngredientsResponse = zod.array(ListIngredientsResponseItem);

/**
 * @summary Add an ingredient to a meal
 */
export const CreateIngredientParams = zod.object({
  id: zod.coerce.number(),
});

export const CreateIngredientBody = zod.object({
  name: zod.string(),
  measurement: zod.string().nullish(),
});

/**
 * @summary Delete an ingredient
 */
export const DeleteIngredientParams = zod.object({
  id: zod.coerce.number(),
  ingredientId: zod.coerce.number(),
});

/**
 * @summary Move a meal to a different day/slot
 */
export const MoveMealParams = zod.object({
  id: zod.coerce.number(),
});

export const MoveMealBody = zod.object({
  scheduledDate: zod.string().date().nullish(),
  mealType: zod.enum(["breakfast", "lunch", "dinner"]).nullish(),
  position: zod.number(),
});

export const MoveMealResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  scheduledDate: zod.string().date().nullish(),
  mealType: zod.enum(["breakfast", "lunch", "dinner"]).nullish(),
  position: zod.number(),
  createdAt: zod.string().datetime({}),
});