# lib/api-zod/src/generated/types/mealWithIngredients.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Ingredient } from "./ingredient";
import type { MealType } from "./mealType";

export interface MealWithIngredients {
  id: number;
  name: string;
  description?: string | null;
  scheduledDate?: string | null;
  mealType?: MealType | null;
  position: number;
  createdAt: string;
  ingredients: Ingredient[];
}